requirements.txt
Pillow>=10.0.0
Flask>=3.0.0
Werkzeug>=3.0.0

